Codecleanups (#4953)
* Minor fixes to get build working

* replace  bulk for loops with addAll

* replace Arrauys List with either List.of or Collections.singletonList

* made blocks static

* misc improvements to collections

* Fixing imports to get build compiling again

* more fixes for impossible != null, unecessary primitive wraps, explicit vararg casts, stream count optimizations

* -Lots of small fixes to unneeded initiations and creations.

* Remove unuecessary redundant type arguments

* Remove redundant casts

* Removed weak types to avoid casting

* removed redundant continues

* -Reverted some casts that were needed.
-Fixed some bad changes.
-Build now works

* maven.compile.source needs to be set, or javadocs defaults to java 1.8

maven.compile.source needs to be set, or javadocs defaults to java 1.8

+++ drools-beliefs/src/main/java/org/drools/beliefs/bayes/EliminationCandidate.java
@@ -42,7 +42,7 @@ public class EliminationCandidate implements Comparable<EliminationCandidate> {
     public void update() {
         // must use the adjacency matrix, and not the vertex.getEdges() for connections,
         // as it gets updated with new connections, during elimination
-        weightRequired = (int) Math.abs(v.getContent().getOutcomes().length);
+        weightRequired = Math.abs(v.getContent().getOutcomes().length);
         newEdgesRequired = 0;
 
         cliqueBitSet = new OpenBitSet(adjMatrix.length);
