+++ drools-examples/src/main/java/org/drools/games/pong/PongUI.java
@@ -21,8 +21,6 @@ import java.awt.event.WindowEvent;
 import java.awt.font.FontRenderContext;
 import java.awt.font.TextLayout;
 
-import javax.swing.*;
-
 import org.drools.games.GameConfiguration;
 import org.drools.games.GameUI;
 import org.kie.api.runtime.KieSession;
@@ -81,7 +79,7 @@ public class PongUI extends GameUI {
         int dash = pconf.getNetDash();
         int x = (tableWidth/2) - (netWidth/2);
         for (int i = 0; i < tableHeight; i = i + dash + gap) {
-            tableG.fillRect( (int) x, i, netWidth, dash );
+            tableG.fillRect(x, i, netWidth, dash);
         }
     }
